backend/requirements.txt
fastapi>=0.103.1,<1.0.0
uvicorn>=0.23.2,<1.0.0
pydantic>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai==0.8.4
httpx>=0.24.1,<1.0.0
python-multipart>=0.0.6,<1.0.0
anyio==4.9.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0  # For database migrations
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.1,<5.0.0
email-validator>=2.0.0,<3.0.0

# Database drivers
psycopg2-binary>=2.9.7,<3.0.0  # For PostgreSQL
# pymysql>=1.1.0,<2.0.0  # For MySQL
# asyncpg>=0.28.0,<1.0.0  # For async PostgreSQL

# Production dependencies
gunicorn>=21.2.0,<22.0.0  # WSGI HTTP Server
uvloop>=0.17.0,<1.0.0  # Fast event loop
httptools>=0.6.0,<1.0.0  # Fast HTTP parsing
